output updated
moved implementation to cpp file for thread and slim_lock, not worth doing so for the others, event is mostly templated, win_exception is fairly small, naive_stack_allocator is templated
